requirements.txt
# Flask Web Framework
Flask==3.1.0
Flask-CORS==4.0.0
Werkzeug==3.1.3

# OCR and Image Processing
pytesseract==0.3.10
easyocr==1.7.2
Pillow==11.3.0
opencv-python==4.11.0.86

# CLIP for Visual Brand Detection (install CLIP separately after this)
torch==2.8.0
torchvision==0.23.0
ftfy==6.3.1
# Run after installing requirements: pip install git+https://github.com/openai/CLIP.git

# Machine Learning
numpy==2.0.2
tensorflow==2.13.0

# Utilities
requests==2.31.0
tqdm==4.67.1
regex==2025.9.1

# Database (built-in with Python)
# sqlite3 - included in Python standard library
